# kapsule-incus-manager/api/schema/openapi.yaml
openapi: "3.1.0"
info:
  title: Kapsule-Incus-Manager API
  version: "0.1.0"
  description: >
    Canonical API definition for the KIM daemon.
    All three frontends (QML, web, CLI) are thin clients against this API.
    The D-Bus interface mirrors every operation defined here.
  license:
    name: GPL-3.0-or-later
servers:
  - url: http://127.0.0.1:8765
    description: Local KIM daemon
...
components:
  schemas:
...
    Operation:
      type: object
      required: [id, description, status, created_at]
      properties:
        id:          { type: string, format: uuid }
        description: { type: string }
        status:
          type: string
          enum: [pending, running, succeeded, failed, cancelled]
        created_at:  { type: string, format: date-time }
        updated_at:  { type: string, format: date-time }
        metadata:    { type: object, additionalProperties: true }
        error:       { type: string }
    InstanceStatus:
      type: string
      enum: [Running, Stopped, Frozen, Error, Unknown]
    InstanceType:
      type: string
      enum: [container, virtual-machine]
    InstanceSummary:
      type: object
      required: [name, type, status, project, remote]
      properties:
        name:       { type: string }
        type:       { $ref: "#/components/schemas/InstanceType" }
        status:     { $ref: "#/components/schemas/InstanceStatus" }
        image:      { type: string }
        project:    { type: string }
        remote:     { type: string }
        created_at: { type: string, format: date-time }
        cpu_usage:  { type: number, description: "CPU usage 0.0–1.0" }
        memory_usage_bytes: { type: integer }
        disk_usage_bytes:   { type: integer }
...
  responses:
...
    Accepted:
      description: Operation accepted and running asynchronously
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Operation" }
...
paths:
...
  /api/v1/provisioning/generic:
    post:
      operationId: createGenericContainer
      summary: Create an incusbox-style container
      tags: [provisioning/generic]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:                { type: string }
                image:               { type: string, default: "images:ubuntu/24.04/cloud" }
                hostname:            { type: string }
                user_name:           { type: string, default: "user" }
                user_uid:            { type: integer, default: 1000 }
                user_gid:            { type: integer, default: 1000 }
                home:                { type: string }
                additional_packages: { type: array, items: { type: string } }
                init_hooks:          { type: array, items: { type: string } }
                pre_init_hooks:      { type: array, items: { type: string } }
                nvidia:              { type: boolean, default: false }
                profiles:            { type: array, items: { type: string } }
                disk_size:           { type: string, default: "20GB" }
                volumes:             { type: array, items: { type: string } }
                project:             { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/generic/{name}/assemble:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    post:
      operationId: assembleGenericContainer
      summary: Run post-create assembly (packages, hooks, shell integration)
      tags: [provisioning/generic]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                packages:          { type: array, items: { type: string } }
                hooks:             { type: array, items: { type: string } }
                shell_integration: { type: boolean, default: true }
                project:           { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
...
  /api/v1/provisioning/generic/{name}/gpus:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    get:
      operationId: listInstanceGpus
      summary: List GPU devices attached to a container
      tags: [provisioning/generic]
      responses:
        "200":
          description: GPU devices
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
    post:
      operationId: attachGpu
      summary: Attach a GPU to a container
      tags: [provisioning/generic]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                dev_name: { type: string, default: "gpu0" }
                gpu_type: { type: string, enum: [physical, mdev, mig, virtio], default: physical }
                pci:      { type: string }
                vendor:   { type: string }
                gid:      { type: integer, default: 44 }
                project:  { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/generic/{name}/gpus/{dev_name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: dev_name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    delete:
      operationId: detachGpu
      summary: Detach a GPU from a container
      tags: [provisioning/generic]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/generic/{name}/usb:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    get:
      operationId: listInstanceUsb
      summary: List USB devices attached to a container
      tags: [provisioning/generic]
      responses:
        "200":
          description: USB devices
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
    post:
      operationId: attachUsb
      summary: Attach a USB device to a container
      tags: [provisioning/generic]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vendor_id, product_id]
              properties:
                vendor_id:  { type: string }
                product_id: { type: string }
                dev_name:   { type: string, default: "usb0" }
                project:    { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/generic/{name}/usb/{dev_name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: dev_name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    delete:
      operationId: detachUsb
      summary: Detach a USB device from a container
      tags: [provisioning/generic]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/generic/{name}/forwards:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    get:
      operationId: listForwards
      summary: List port-forward proxy devices on a container
      tags: [provisioning/generic]
      responses:
        "200":
          description: Proxy devices
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
    post:
      operationId: addForward
      summary: Add a port-forward proxy device
      tags: [provisioning/generic]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host_port, guest_port]
              properties:
                host_port:   { type: integer }
                guest_port:  { type: integer }
                protocol:    { type: string, default: "tcp" }
                listen_addr: { type: string, default: "127.0.0.1" }
                dev_name:    { type: string }
                project:     { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/generic/{name}/forwards/{dev_name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: dev_name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    delete:
      operationId: removeForward
      summary: Remove a port-forward proxy device
      tags: [provisioning/generic]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/generic/{name}/snapshots:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    get:
      operationId: listGenericSnapshots
      summary: List snapshots for a container
      tags: [provisioning/generic]
      responses:
        "200":
          description: Snapshot list
          content:
            application/json:
              schema: { type: array, items: { type: object, additionalProperties: true } }
    post:
      operationId: createGenericSnapshot
      summary: Create a snapshot
      tags: [provisioning/generic]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [snapshot]
              properties:
                snapshot: { type: string }
                stateful: { type: boolean, default: false }
                project:  { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/generic/{name}/snapshots/schedule:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    put:
      operationId: setGenericSnapshotSchedule
      summary: Set automatic snapshot schedule
      tags: [provisioning/generic]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [schedule]
              properties:
                schedule: { type: string }
                expiry:   { type: string }
                project:  { type: string, default: "" }
      responses:
        "200":
          description: Schedule updated
    delete:
      operationId: disableGenericSnapshotSchedule
      summary: Disable automatic snapshots
      tags: [provisioning/generic]
      responses:
        "200":
          description: Schedule disabled
  /api/v1/provisioning/generic/{name}/snapshots/{snapshot}/restore:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: snapshot, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    post:
      operationId: restoreGenericSnapshot
      summary: Restore a snapshot
      tags: [provisioning/generic]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/generic/{name}/snapshots/{snapshot}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
      - { name: snapshot, in: path, required: true, schema: { type: string } }
      - { name: project, in: query, schema: { type: string, default: "" } }
    delete:
      operationId: deleteGenericSnapshot
      summary: Delete a snapshot
      tags: [provisioning/generic]
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/generic/fleet:
    get:
      operationId: listGenericFleet
      summary: List all generic containers
      tags: [provisioning/generic]
      parameters:
        - { name: project, in: query, schema: { type: string, default: "" } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: Container list
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/InstanceSummary" } }
  /api/v1/provisioning/generic/fleet/start:
    post:
      operationId: fleetStartGeneric
      summary: Start multiple containers
      tags: [provisioning/generic]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [names]
              properties:
                names:   { type: array, items: { type: string } }
                project: { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/generic/fleet/stop:
    post:
      operationId: fleetStopGeneric
      summary: Stop multiple containers
      tags: [provisioning/generic]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [names]
              properties:
                names:   { type: array, items: { type: string } }
                project: { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
  /api/v1/provisioning/generic/publish:
    post:
      operationId: publishGenericContainer
      summary: Publish a container as a reusable Incus image
      tags: [provisioning/generic]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string }
                alias:       { type: string }
                description: { type: string }
                public:      { type: boolean, default: false }
                project:     { type: string, default: "" }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }


  # ── Waydroid container provisioning ──────────────────────────────────────
...